+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Main (unreleased)
 
 - Agent Management: Introduce support for templated configuration. (@jcreixell)
 
+### Enhancements
+ 
+- Flow Windows service: Support environment variables. (@jkroepke)
+
 ### Bugfixes
 
 - Permit `X-Faro-Session-ID` header in CORS requests for the `faro.receiver`